revert the app to use native js instead of express

+++ controllers/delete-controller.js
@@ -1,16 +1,21 @@
 const MySqlService = require('../services/mysqlService');
 const sendResponse = require('../utils/server-utils/serverResponse');
 const queryBuilder = require('../utils/query-builder/queryBuilder');
+const urls = require('../utils/constants/urls');
 
 module.exports = (req, res) => {
-  const mysqlApi = new MySqlService();
-  let userId = req.params.id;
-  let namesQuery = queryBuilder.deleteDataQuery(userId);
+  if (req.path.startsWith(urls.DELETE)) {
+    const mysqlApi = new MySqlService();
+    let userId = req.url.split('/').pop();
+    let namesQuery = queryBuilder.deleteDataQuery(userId);
 
-  return mysqlApi
-    .execute(namesQuery)
-    .then(data => {
-      return sendResponse(res, data);
-    })
-    .catch(e => sendResponse(res, e.message));
+    return mysqlApi
+      .execute(namesQuery)
+      .then(data => {
+        return sendResponse(res, data);
+      })
+      .catch(e => sendResponse(res, e.message));
+  } else {
+    return true;
+  }
 };
